bot-telegram/pyproject.toml
[project]
name = "bot-telegram"
version = "0.1.0"
description = "MQTT subscriber that forwards ERROR events to Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
bot-telegram = "bot_telegram.presentation.main:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
